+++ cli/tests/pcluster/validators/test_networking_validators.py
@@ -10,7 +10,10 @@
 # limitations under the License.
 import os
 
-from pcluster.validators.networking_validators import SecurityGroupsValidator, SubnetsValidator
+import pytest
+
+from pcluster.config.cluster_config import SlurmComputeResource, SlurmQueue, SlurmQueueNetworking
+from pcluster.validators.networking_validators import SecurityGroupsValidator, SingleSubnetValidator, SubnetsValidator
 from tests.pcluster.aws.dummy_aws_api import mock_aws_api
 from tests.pcluster.validators.utils import assert_failure_messages
 
@@ -63,3 +66,64 @@ def test_ec2_subnet_id_validator(mocker):
     # TODO test with invalid key
     actual_failures = SubnetsValidator().execute(["subnet-12345678", "subnet-23456789"])
     assert_failure_messages(actual_failures, None)
+
+
+@pytest.mark.parametrize(
+    "queues, failure_message",
+    [
+        (
+            [
+                SlurmQueue(
+                    name="queue",
+                    networking=SlurmQueueNetworking(subnet_ids=["subnet-11111111"]),
+                    compute_resources=[
+                        SlurmComputeResource(
+                            instance_type="test",
+                            name="test1",
+                        )
+                    ],
+                ),
+                SlurmQueue(
+                    name="queue",
+                    networking=SlurmQueueNetworking(subnet_ids=["subnet-00000000"]),
+                    compute_resources=[
+                        SlurmComputeResource(
+                            instance_type="test",
+                            name="test1",
+                        )
+                    ],
+                ),
+            ],
+            "The SubnetId used for all of the queues should be the same",
+        ),
+        (
+            [
+                SlurmQueue(
+                    name="queue",
+                    networking=SlurmQueueNetworking(subnet_ids=["subnet-00000000"]),
+                    compute_resources=[
+                        SlurmComputeResource(
+                            instance_type="test",
+                            name="test1",
+                        )
+                    ],
+                ),
+                SlurmQueue(
+                    name="queue",
+                    networking=SlurmQueueNetworking(subnet_ids=["subnet-00000000"]),
+                    compute_resources=[
+                        SlurmComputeResource(
+                            instance_type="test",
+                            name="test1",
+                        )
+                    ],
+                ),
+            ],
+            None,
+        ),
+    ],
+)
+def test_single_subnet_validator(queues, failure_message):
+    actual_failure = SingleSubnetValidator().execute(queues)
+
+    assert_failure_messages(actual_failure, failure_message)